LayoutTests/imported/w3c/web-platform-tests/uievents/mouse/mouse_buttons_back_forward-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Test Description: Tests that the mouseup event is prevented.
Click the back mouse button
Click the back mouse forward


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN Tests that the mouseup is preventable.
